[update]: 때때로 모달 안열리는 문제의 원인으로 의심되는 함수 위치 이동
